refactor: 예약 컴포넌트를 components/common/에서 components/reservation/으로 분리
